Update Category: - display order
Type:
- deposit -> int

District Service:
- update entity not found exception

+++ src/main/java/org/avengers/capstone/hostelrenting/model/Type.java
@@ -44,7 +44,7 @@ public class Type {
 
     // unit: month
     @Column(nullable = false)
-    private float deposit;
+    private int deposit;
 
     @Transient
     private int schoolmate;
